Update changelog URL
Link to the latest copy of the changelog document, rather than release page.

There are no changelog notes for terraform-ls GitHub releases, so link directly to the changelog file instead.

Since this plugin is only used to install the latest terraform-ls version, link to the main branch `CHANGELOG.md`, rather than the version as of the appropriate tag. Sometimes the changelog is modified after the tag with forgotten information, and we don't want to hide that from users.