(GH-628) Add semver based version pin for Language Server
Previously the Language Server would always update to the latest version however
there are scenarios where using an older version is required.  This commit adds
a new `version` configuration setting which is a semver range (as per the
@hashicorp/js-releases module). This setting is then used to find an appropriate
release when calling the Hashi release API.

This commit also adds validation on the `version` setting to ensure it is a
valid string and warns the user of the potentially bad setting.

This commit then updates the Language Server installer to not only upgrade the
server, but also downgrade it.