fix: launch LS even if path contains escapable characters
This fixes a long undiscovered bug which would cause the LS to never launch
in situation where there is an escapable character (such as whitespace)
in the *default* path leading to the LS binary.

The root cause is string concatenation of full commands (and
`child_process.exec` API expecteding already escaped string)
when checking LS version, instead of passing arguments individually.

This problem would also never be surfaced to the user due to some other
issues with error reporting (that can be addressed in a separate PR).