Extract logic for determining LS paths
Path to either the LS binary or to the installation folder
is needed in multiple places and sadly isn't as trivial.
This extraction ensures that the same mistakes (such as
forgetting to reflect Windows exe extensions) aren't repeated.

Windows would normally still find and execute a binary
even without the "exe" extension, but some further checks
(testing file existence/permissions) will be added
in a separate PR and that requires knowing the real binary name.