Improve error handling (#691)
* surface any exceptions from LS update as error popup

* make language client name & ID more descriptive

This helps in situations where either of those is used
in an error message (raised by upstream LSP library).

* provide more helpful error when LS binary isn't accessible

The upstream library doesn't perform any checks on the path before
passing it to `child_process.spawn()`. To make matters worse
when the spawning fails the library resurfaces that error
as a very generic error "Couldn't start client <client name>".