(maint) Set syntax and format style guidelines
Update eslint and prettier packages and add style and format guidelines to editorconfig and VS Code worskpace settings file.

This aligns to company standards for code style and formatting and ensures consistent spacing, line endings, and file encodings. This does not enable any new eslint rules, as that is out of scope for this line of work.

This also configures VS Code to use prettier and eslint for all typescript, javascript and json formatting and use eslint and prettier to enforce guidelines.