Improve first use experience
Add a check to see if the current workspace has been initialized. If not, pop a warning message informing the user why that is important for the performance of the extension and provide a set of actions for the user to choose. They can either ignore the message, see documentation, or choose to execute terraform init.

When choosing terraform init, it will open the command pallete and wait for the user to execute. This accomplishes the init task as well as teaching the user how the builtin commands help.