github: Update publishing workflow & make it single-step
Changelog in Marketplace entry is sourced automatically from Changelog.md
We can continue updating that changelog manually.

We can also continue copy-pasting the relevant part of changelog
into GitHub Release notes.

We can use Linux instead of Windows since Windows is not a requirement for vsce.

Lastly we de-ephasise the vsix package as an artifact which can still be found
and used but ideally we want folks to install from the Marketplace.
This is a compromise between stopping publishing of vsix entirely vs
automatically creating a release to which we'd upload it as an asset.