Update minimum VS Code engine to 1.52.0 (#780)
This updates the vscode-languageclient library to the 7.0.0 release, which requires the minimum vs code engine to be 1.52.0.

The vscode-languageclient 7.0.0 release is the current supported release, but we pinned to 7.0.0-next.12. We shouldn't be running on the interim releases, but we were waiting on https://github.com/hashicorp/terraform-ls/pull/608 to merge to support the changed jsonrpc constructs in 7.0.0. Now that that is merged, released, and in the wild for a significant amount of time we can move to the 7.0.0 vscode-languageclient version.

This requires us to increment the minimum supported version of VS Code to 1.52.0. This means any user with a VS Code version lower than 1.52.0 will not see an updated extension in the marketplace, but will still receive updated terraform-ls releases through the automatic updater.