Silent Langauge Server Installation
This uses the existing VS Code update settings to determine whether to check for updated LS and/or whether to update the LS server without prompting the user.

A user that sets `extensions.autoCheckUpdates` to true will automatically check for updates to the LS. If set to false, the extension will not check for updates to the LS.

A user that sets `extensions.autoUpdate` to true will receive new versions of the LS without a prompt.

The installation progress message is moved to the `Window` position to eliminate the final distraction but maintain notification of what is going on.

The end result is that if a user has the default settings for `extensions.autoCheckUpdates` and for `extensions.autoUpdate`, the LS will be kept up to date without prompting or distracting the user.