Refactor to Single LanguageClient
Refactors ClientHandler to only use one `LanguageClient` per extension activation. Removes the Map of `LanguageClients` and ensures that any attempt to use a `LanguageClient` or method happens after the onReady event finishes.

Previously the VS Code extension would create a `LanguageClient` for each folder that was added to the current workspace. Each `LanguageClient` would handle all events and receive notifications/requests from the LS for the folders it was responsible for.

This was handled by a `Map` of clients created at activation and populated each time a client was created. However there were conditions in which an event would occur before a client was created or finished being ready, so an attempt to access the `Map` would result in an unhandled exception. A common example was having an file that was already opened when VS Code was opened. The file would trigger an `onDidChangeVisibleTextEditors` event, which would try to get the list of module calls used, which would fail because the client had not started yet.

The terraform-ls sever implemented multi-folder/workspace support in https://github.com/hashicorp/terraform-ls/pull/502. This means the VS Code extension no longer needs multiple clients to track how many folders are added to the current workspace and which event for which file needs to be sent at certain times.