Simplify creating TerraformLanguageClient
Refactor the createTerraformClient method to reduce complexity and verbosity by using language constructs. Move variables closer to usage for future method extraction.

In d707a6d we moved to a single LanguageClient in the extension. This means we no longer need to track the `location` or `workspaceFolder` to pull settings or display information per folder. This greatly simplifies the code accessing settings and creating the LanguageClient.