Extract supported commands to class property
Moves the list of supported commands to a private property of the ClientHandler class and populates it when the client and LS are initialized.

Previously you could attempt to ask the server what commands were supported before the client and server had finished the initialization process. This would result in an unhandled exception and leave the client in an unknown state.

Moving the query to the `onReady` function ensures we only ask after the server has finished initializing and is ready to respond.