Enable LS in main activation flow
Directly call `updateLanguageServer` in the main execution flow and await return instead of inside a `executeCommand` call.

The `terraform.enableLanguageServer` command checks if the LS is enabled, then returns `updateLanguageServer`. We already checked if the LS is enabled before calling `terraform.enableLanguageServer`, so we are duplicating effort here. There is also no reason to nest the async calls to the main functionality of our extension inside a `executeCommand` call.

Technically `executeCommand` calls are supposed to be commands that initiate actions/commands/things inside the workspace, not long running operations. This does not match well with running the primary execution work flow of our extension.