Brand OutputChannel
Move creation of the OutputChannel to the main flow and only create one outputChannel for the entire extension. Use 'HashiCorp Terraform' as the name the OutputChannel to brand our extension.

The OutputChannel is the main way to log information for the user to see. By convention there is one for each extension and the content is meant to be human readable. The `name` is displayed in a drop down in the `Output` pane in the editor, so it should be reconizable and readable.

Currently we set `terraform-ls` or `terraform-ls 'location'` for the name. This lower cased name is hard to find in the drop down, espcially compared to other extensions that use their company name and language to disambiguate from others.

Setting the name to `HashiCorp Terraform` ensures we are easy to read and differentiate from other extensions. Creating only one for the entire extension makes it easier for users to find out information as well as allowing us to automate collecting logs for bug report submission.