Use vscode.workspace.fs instead of fs package (#852)
Replaces all usages inside extension of the fs package with the `vscode.workspace.fs` wrapper. This ensures safe usage of the filesystem and allows us in the future to move to remote extension use.

Fixes bug where zip is not deleted after extraction. This also helps in the future when we minify the extension and del does not work.