Fix Detecting Terraform files in status calls
This modifies the moduleCalls and moduleProviders views and the updateTerraformStatusBar function to detect if the open file is an `Untitled` file or other type of file and return early instead of trying to determine Terraform module information.

Previously the editor would attempt to determine terraform module status whenever a file was opened. This worked well if a Terraform file was opened, but would throw exceptions if a file was opened that did not contain Terraform code. This includes `Untitled` files which are open files that are not saved yet and don't exist on the filesystem yet. The fix is to ignore any non Terraform file.

This contains a naive check to determine if a file is a Terraform file, which checks the file extension for `tf`. This suits for now but possibly might need to be expanded in the future if the extension supports more Terraform file types.