Silent Staged LS Installer (#868)
* Silent Staged Installer

The terraform-ls installer no longer prompts the user for whether to install an update (or downgrade) and now respects the `extensions.autoUpdate` and `extensions.autoCheckUpdates` settings.

If `extensions.autoUpdate` is set to `true`, the installer checks for updates on activation and installs a new version if available. If set to `false`, then it does not check for updates.

When installing an updated terraform-ls version, it installs to a staging directory and continues activation. On restart it will move the terraform-ls binary from the staging location to the extension path and the new version will be used.

If `extensions.autoCheckUpdates` is set to `true`, the installer automatically checks for updates every 24 hours. If set to `false`, it does not check on a schedule.

An update can either be an upgrade or downgrade, depending on the version set in terraform.languageServer.requiredVersion. If latest or no version is specified, the most current terraform-ls version is used. In either an upgrade or a downgrade, the installer will perform the required action without asking the user.