Fix creating process with environment variables
This ensures the child process we create for the Language Server contains the process environment variables needed to find other binaries on the host.

When creating the process for the Language Server, we relied on process.spawn to copy the environment variables to the child process. This works in most cases, but fails to contain the host environment variables in certain scenarios. For example, opening VS Code from the macOS Launcher. This leaves our process with important environment variables like PATH, and results in failures down the line attempting to use other binaries like terraform.

These errors are mostly silent, leaving the user confused why something isn't working. It's also hard to debug, because the correct variables (especially PATH) are present and work outside our app.