Update eslint and minimal ruleset
This updates eslint and related plugins to latest package versions. It also brings the eslint config into line with default config published by the VS Code Extension yo template.

Most of the code is compliant with the eslint rulsets, with the exception of naming convention for fields and variables. Where we could update the naming, we did. Where we could not eslint disable lines were added.