Deprecate terraform.languageServer.requiredVersion
This adds a deprecation message to the `terraform.languageServer.requiredVersion` setting.

This setting will be removed in a future release when this extension adopts platform specific extension publishing.

In a future release, the extension will bundle the terraform-ls binary inside the published VSIX. This eliminates the need to download the binary at runtime, but also means the user cannot change the terraform-ls version used. Since they cannot change the version, this setting will no longer be needed.

Users who wish to use a different version than the one shipped with the extension will use `terraform.languageServer.pathToBinary` instead and install the terraform-ls binary using one of the release artifacts or through a package manager.

More detailed documentation on this process will be added when platform specific extension support is added.