Download terraform-ls script
This adds a typescript file that downloads a version of terraform-ls that matches the current OS and architecture of the machine running the script. The version of terraform-ls that is chosen is pulled from the `langServer` attribute in the `package.json` file.

This script can be invoked several ways:

```
npm run download:ls

npm run download:ls --target=darwin-x64
```

It can also be combined with `npm package` to download and package an alternate OS and architecture:

```
ls_target=linux-x64 npm run package -- --target=linux-x64

$env:ls_target="linux-x64"; npm run package -- --target=linux-x64
```