Publish Platform Specific Extension (#905)
This converts the extension to use a bundled terraform-ls binary and packages the VSIX as a Platform Specific Extension. A matrix of all supported platforms and architectures run the commands necessary to bundle terraform-ls and the extension into a VSIX. After packaging, it publishes all stored VSIXs to the Marketplace. This results in a VSIX that is ready to run at install time, instead of having to use the network at runtime to download a terraform-ls binary.

This also adds minification as a final step to the interactive development and packaging processes. This now produces a single file extension, compared to previously shipping several thousand files. This results in a much faster installation experience and reduces the deployment size of the extension.

Implementation

* Download terraform-ls script

This adds a typescript file that downloads a version of terraform-ls that matches the current OS and architecture of the machine running the script **or** using the `--target` parameter **or**using an environment variable `ls_target`. The version of terraform-ls that is chosen is pulled from the `langServer` attribute in the `package.json` file.

This script can be invoked several ways:

```
npm run download:ls

ls_target=darwin-x64 npm run download:ls

npm run download:ls --target=darwin-x64
```

It can also be combined with `npm package` to download and package an alternate OS and architecture:

```
ls_target=linux-x64 npm run package -- --target=linux-x64

$env:ls_target="linux-x64"; npm run package -- --target=linux-x64
```

* Use bundled terraform-ls

This converts the extension to use the bundled terraform-ls binary instead of one downloaded at runtime. This effectively changes the extension to not download the terraform-ls binary and instead rely on one being present in the `bin` directory inside the extension path.

It also moves @hashicorp/js-releases to a development dependency, which allows openpgp, semver, and yauzl to be removed. These dependencies are no longer needed as runtime dependencies in the extension as we no longer download the terraform-ls bianry at runtime.

We still will use @hashicorp/js-releases to download the terraform-ls binary at development and package time.

* Update github actions for platform specific vsix

This updates the publish and preview github actions to package platform specific extensions.

A matrix of all supported platforms and architectures run the commands necessary to bundle terraform-ls and the extension into a VSIX. After packaging, it publishes all stored VSIXs to the Marketplace.

* Minify the extension

This adds esbuild as a final step to the interactive development and packaging processes. This produces a single file extension for deployment, compared to several thousand files.

This results in a much faster installation experience and reduces the deployment size of the extension.