Standardize Changelog (#909)
The current changelog contains a list of all versions back to the inception of the extension and is an amalgam of different processes for classifying and sorting different issues or pull requests added to the log.

This PR moves all releases 1.4.0 and older to a new Previous Releases section, to reduce the length of the log while maintaining references to those versions. For each release header and classification, this adds dates and normalizes the capitalization, naming conventions, and ordering. Lastly it adds links to the tag comparison for each release, so a view of changes for each release can be seen.

This results in a human readable and accessible format that works both in GitHub and in the VS Code Marketplace.