Add Terraform language tmgrammar snapshots
This adds a snapshot file for every language construct from https://www.terraform.io/language that makes sense to test syntax grammar for. Some constructs are left out because it does not make sense to test grammar for it now.

A minimal listing of numeric and string functions are added to expressions_functions.tf, which is not the complete listing of functions. The existing grammar uses a regex with a list of function names, so this small list is enough to test at the moment. If we decide to change the regex for functions in the future, we should add more test cases to this file.