Fix Nested Map Highlighting (#925)
* Add a base test case for nested maps

This adds a terraform file containing nested maps and a snapshot
test capturing the broken highlighting

* fix highlighting of nested maps

This commit adds `objects` as an include to `objects`. It allows
deeply nested objects and results in a new `meta.braces.terraform`
scope for each nested level.

* Further improve highlighting of nested maps

This changes the scope for object keys from `string.unquoted.terraform`
to `variable.other.readwrite.terraform`, allowing them to be highlighted
in line with other variables.

With the update of the `keyword.operator` the `=` inside nested maps is
also highlighted in line with variable assignments. This change also
applied to inline for expressions, so we had to extend the pattern
for matching `=>`.

Matching `=>` here is only a quick-fix. When we improve the detection
of inline `for` loops, it can be removed.