Add new scope to terraform block type and name
This changes the existing block matcher to differentiate the type and name of a given block.

Previously the type and name of a block were scoped as `string.quoted.double`. This change scopes them to an `entity.name.tag`, which is commonly used to indicate a tag or attribute.

This will match Terraform blocks like `resource "aws_instance" "web" {` or `module "foo" {`