Resolve issue with tfvars comment toggling (#937)
This fixes tfvar file detection by adding the tfvar language and file extension to the activationEvents array. This ensures that even if a folder is opened with only tfvar files in it, our extension is activated.

It also adds the language-configuration file to the tfvars language contribution section, which enables language features like comment toggling. 

Co-authored-by: James Pogran <jpogran@outlook.com>