syntax: simplify capturing of block labels
HLC considers the quotes to be effectively part of the same token (label), i.e. it treats both quoted or unqouted labels the same, although quoted labels are considered best practice.

We reflect the equivalence here and assume that quotes (if present) will be highlighted the same as the string between the quotes.