Improve Extension Documentation (#942)
This commit updates the README with more user focused documentation about extension usage, features, troubleshooting info, an improved quick start, and more detailed workflow scenarios. The aim is to be as comprehensive and helpful as possible without going into dedicated website territory.

To do this required significantly changing the order and content of the sections. This allows us to have linkable headers that we can provide in tickets and when interacting with our community.

It also adds example screenshots of features in use. These are purposely ignored by the packaging command to save space in the VSIX. They are referenced using Markdown links that will be automatically translated by vsce when packaging, so will be available to internet connected clients when viewed inside the Extension pane. If there is no internet, they won't be resolvable, but that is a small price to pay for a small install size.

This also includes corrected and/or updated information that replaces the following Wiki pages:

https://github.com/hashicorp/vscode-terraform/wiki/Install-a-Pre-transfer-Version
https://github.com/hashicorp/vscode-terraform/wiki/Manually-Setting-the-Terraform-Executable-Path

Co-authored-by: Radek Simko <radek.simko@gmail.com>
Co-authored-by: Daniel Banck <dbanck@users.noreply.github.com>