Enable preview workflow to download terraform-ls preview versions (#954)
This updates downloader.ts to download a pre-release version of terraform-ls on build instead of just stable releases. It draws the version to download from the `langServer` and `preview` package.json fields.

This it sets the version of terraform-ls in package.json to a version specified in an environment variable, if present. If not present, it uses the existing value.

Finally it adds a workflow input to the preview GHA that accepts a terraform-ls version to download.