Organize extension settings into Sections (#1024)
This groups and orders the existing settings into VS Code Setting sections while preserving the exact keys for each setting.

The extension previously utilized different prefixes to indicate different areas like terraform-ls.rootModules or terraform-ls.terraformExecTimeout. This leaves settings in a seemingly random order in the Settings UI and makes it hard to autocomplete inside the json file.

We solve this by organizing the extension settings into sections in the contribute part of the package.json manifest. These sections provide a logical order to extension settings, especially if the extension has many settings or many settings that are nested. This also allows us to separate out configuration options, like a separate section for Experimental Options.