Remove terraform.languageServer.trace.server (#1048)
This removes the terraform.languageServer.trace.server setting. This setting is not used by the extension or terraform-ls.

Removing it entirely does not break anything, it will just be grayed out in the JSON file and won't show at all in the UI. If a user hovers over the setting in the JSON file, they'll be given a general message that the setting is no longer used.