Indicate Language Server Status
This uses the new vscode.languages.createLanguageStatusItem API to add
a status indicator next to the language ID in the status bar. This
allows us to show the current state of the language server as it goes
from starting to running to stopping, as well as any error conditions.

It also gives the ability to hook commands to the indicator, so we can
show a menu to enable, disable, restart, or show logs.