Github Release Notes Generator file (#1051)
This adds a configuration file to the repo for the GitHub Release Notes
API to read. This does not create a release or any other change to the
repo, it just generates a changelog that is very close to how we
currently format the changelog. This aids in reconciling changes before
release time.