Optimize main entry point execution path
This extracts functions, re-orders execution and otherwise cleans up the main entry point for the extension.

Migrating settings is extracted to a function to prepare for the work to be done renaming settings later on.

The command registration calls are reorganized to prioritize registering first those that do not require the terraform-ls to function. Starting the terraform-ls binary is then called directly before registering the rest of the commands and providers.

The end result is registering and activating the parts that need the LS only if it is enabled.