Remove command prefix (#1075)
This removes generating a command prefix when registering commands for the language client connecting to terraform-ls.

Previously the extension would create a LanguageClient/terraform-ls pair per 'folder' in a workspace. This means there was a terraform-ls process per folder to keep track of. The https://github.com/hashicorp/terraform-ls/pull/502 PR enabled a single terraform-ls process to handle multi-folder workspaces, which meant we could refactor the extension in https://github.com/hashicorp/vscode-terraform/pull/845 to use a single LanguageClient instance. A leftover from the original design was a prefix for commands registered between terraform-ls and the client. This ensured unique command IDs when multiple clients were present, but is no longer necessary with a single language client.