Extract LanguageClient from ClientHandler (#1082)
This extracts the LanguageClient creation from ClientHandler and puts it directly inside the activation point.

In order to implement https://github.com/hashicorp/terraform-ls/issues/722 we need to add a new StaticFeature to register a client-side command for terraform-ls to call when it knows the `terraform.providers` view needs to be refreshed.

StaticFeatures are registered using the LanguageClient.registerFeature method, which means the ClientHandler class needs to create the StaticFeature. The StaticFeature needs both the LanguageClient and `terraform.providers` view created before it can be initialized. The LanguageClient is created by the ClientHandler class. This all results in a cyclic dependency.

We could resolve this cyclic dependency by adding more responsibility to ClientHandler, but this introduces more complexity to the class. This will become increasingly hard to deal with as more aspects like StaticFeature are added.

We resolve this by extracting the creation of LanguageClient and moving dependent features like the StaticFeatures to the main activation method. This puts all the parts that rely on each other in the same place where the data needed is located to make decisions about whether they are activated or not.

This builds on work done in:

- https://github.com/hashicorp/vscode-terraform/pull/1073
- https://github.com/hashicorp/vscode-terraform/pull/1074
- https://github.com/hashicorp/vscode-terraform/pull/1075
- https://github.com/hashicorp/vscode-terraform/pull/1079