Automatically refresh Terraform Providers View
This commit automatically refreshes the Terraform Providers View when providers change in open document.

This introduces a StaticFeature which registers a handler for the `client.refreshModuleProviders` request which terraform-ls can call to refresh the data shown in the Terraform Providers view. The terraform-ls server was modified to send `client.refreshModuleProviders` requests in https://github.com/hashicorp/terraform-ls/pull/902