Automatically refresh Terraform Module Calls View
This commit automatically refreshes the Terraform Module Calls View when providers change in open document.

This introduces a StaticFeature which registers a handler for the `client.refreshModuleCalls` request which terraform-ls can call to refresh the data shown in the Terraform Module Calls view. The terraform-ls server was modified to send `client.refreshModuleCalls` requests in hashicorp/terraform-ls#902