Fix enabling or disabling language server commands
This updates the method the commands to enable or disable the language server use to modify the extension configuration. Previously this relied on untyped calls which are not allowed anymore for Object.assign.

This also fixes a scoping problem with the original code. It previously read the current setting value, but only updated the Global setting value. If a user set the value in a Workspace setting, this would update the Global setting and miss updating the Workspace setting. This now detects where the setting was set, and updates the appropriate placement.

This also directly uses the config method to determine if the language server is enabled or not. Previously it used the `enabled()` function which provided a default value of false if the setting was unable to be read. This is more a syntax/preference than a strict need. I think the intention is better reflected by writing it out here and removes ambiguity.

This will allow https://github.com/hashicorp/vscode-terraform/pull/1117 to be merged.