Fix enabling or disabling language server commands (#1131)
This updates the method the commands to enable or disable the language server use to modify the extension configuration. Previously this relied on untyped calls which are not allowed anymore for this method.

The root of this problem is that we made an untyped JSON object out of `languageServer` setting, instead of having separate fields. This has caused/is causing problems everywhere we try to access and change these setting values. Since it's treated like one object we can't set one value (external), we have to set all the values at the same time. In short, the `vscode.WorkspaceConfiguration.update()` method operates on `terraform.languageServer` not `terraform.languageServer.external`. To complicate it further, these settings can be different depending on scope (user, workspace, folder, etc).

To handle this previously we pulled the existing setting values (albeit incompletely) and used `Object.assign` to map the new value onto the existing values. This now does not work because `Object.assign` does not accept any as an object to work on anymore. We have to provide a typed object.

To fix this we we provide the small `LanguageServerSettings` interface to the `vscode.WorkspaceConfiguration.get()` method which now returns typed object that we can set values to, then pass to `vscode.WorkspaceConfiguration.update()` and have it work.

One last thing to consider is scope. We read the value several times before we assign it. VS Code automatically figures out the value from the scopes when we call get, so we don't have to specify scope. We do have to specify scope when updating the settings. We get scope by using the `vscode.WorkspaceConfiguration.inspect()` method, which gives us the information on which scope to update.

This will be all avoided when we do #791, which will allow us to only update one value at a time.

This also directly uses the config method to determine if the language server is enabled or not. Previously it used the `enabled()` function which provided a default value of false if the setting was unable to be read. This is more a syntax/preference than a strict need. I think the intention is better reflected by writing it out here and removes ambiguity. 

This will allow #1117 to be merged.