Set proper scope for machine based settings (#1164)
This sets the following settings to `machine-overridable` scope. This ensures these settings will be read in the correct location whether the user is using the extension locally or in a remote context like the Remote Extensions or Codespaces.

It is set to `machine-overridable` instead of just `machine`, to still allow users to have these settings set inside a Workspace file or in the Workspace folder. If we set it to just `machine`, the settings would not be modifiable in the Workspace file or in the Workspace folder scopes.

```
terraform.languageServer.path
terraform.languageServer.args
terraform.languageServer.terraform.path
terraform.languageServer.terraform.timeout
terraform.languageServer.terraform.logFilePath
terraform.languageServer.rootModules
terraform.languageServer.excludeRootModules
terraform.languageServer.ignoreDirectoryNames
```