Terraform View container
Create a new Terraform View Container that will house the existing terraform.providers and terraform.modules views. These views will no longer be inside the Explorer View Container. This Terraform View Container will support all the built-in mechanisms users expect out of views, most notable of which is the ability to move it to other places in the editor like the new SideBar or be able to be hidden altogether. This enables us to have a single place to land new views as well as allow the user to customize as they see fit.